clean up integration test
